+++ server/app/__init__.py
@@ -1,0 +1,1 @@
+"""AngysGuard managed-test server package."""